requirements.txt
altgraph==0.17.4
boto3==1.35.65
botocore==1.35.65
jmespath==1.0.1
macholib==1.16.3
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
setuptools==75.6.0
six==1.16.0
urllib3==2.2.3
